working with testing,animation

+++ PlanRaderTask/Utlits/Extentions/Foundation+Extentions/Date+Extention .swift
@@ -1,9 +1,9 @@
 //
 //  UIBundle+Extention.swift
-//  Fleez
+//  PlanRaderTask
 //
-//  Created by A One Way To Allah on 10/10/21.
-//  Copyright © 2021 A One Way To Allah. All rights reserved.
+//  Created by A One Way To Allah on 05/03/2022.
+
 //
 
 import Foundation
@@ -14,7 +14,7 @@ extension Date{
     var getTimeStringFromDate : String {
         let dateFormatter = DateFormatter()
         dateFormatter.dateStyle = .medium
-        dateFormatter.timeStyle = .long
+        dateFormatter.timeStyle = .short
         dateFormatter.locale=Locale(identifier: "en")
         let dataString = dateFormatter.string(from: self)
         return dataString
